Add memcache support to the file_metadata REST terminus.
This clones the (empty) REST file_metadata terminus class, and extends it to
have a memcache look-aside for data in the cache.  This is an optimistic
cache: it will assume that if data is in the cache it is correct, and will sit
on it forever; no rechecking is done at this time.

This builds a model that saves the round-trip to the server for metadata, but
doesn't solve the expiration and refresh problem.  Those are assumed to be
addressed externally, which is harder than it sounds. :/